+++ src/helpers.ts
@@ -1,5 +1,5 @@
 import { fromByteArray } from 'base64-js';
-import { Layer, Color, BlendMode, LayerColor } from './psd';
+import { Layer, BlendMode, LayerColor } from './psd';
 
 export const fromBlendMode: { [key: string]: string } = {};
 export const toBlendMode: { [key: string]: BlendMode } = {
@@ -282,90 +282,6 @@ export function writeDataRLE(buffer: Uint8Array, { data }: PixelData, width: num
 	}
 
 	return buffer.slice(0, o);
-}
-
-// h = <0, 360>; s, v = <0, 100>
-export function hsv2rgb(h: number, s: number, v: number): Color {
-	h = Math.max(0, Math.min(360, h === 360 ? 0 : h));
-	s = Math.max(0, Math.min(1, s / 100));
-	v = Math.max(0, Math.min(1, v / 100));
-
-	let r = v;
-	let g = v;
-	let b = v;
-
-	if (s !== 0) {
-		h /= 60;
-		const i = Math.floor(h);
-		const f = h - i;
-		const p = v * (1 - s);
-		const q = v * (1 - s * f);
-		const t = v * (1 - s * (1 - f));
-
-		switch (i) {
-			case 0:
-				r = v;
-				g = t;
-				b = p;
-				break;
-			case 1:
-				r = q;
-				g = v;
-				b = p;
-				break;
-			case 2:
-				r = p;
-				g = v;
-				b = t;
-				break;
-			case 3:
-				r = p;
-				g = q;
-				b = v;
-				break;
-			case 4:
-				r = t;
-				g = p;
-				b = v;
-				break;
-			default:
-				r = v;
-				g = p;
-				b = q;
-		}
-	}
-
-	return [
-		r * 255,
-		g * 255,
-		b * 255,
-		255,
-	];
-}
-
-export function lab2rgb(l: number, a: number, bb: number, alpha = 255) {
-	let y = (l + 16) / 116;
-	let x = a / 500 + y;
-	let z = y - bb / 200;
-
-	x = 0.95047 * ((x * x * x > 0.008856) ? x * x * x : (x - 16 / 116) / 7.787);
-	y = 1.00000 * ((y * y * y > 0.008856) ? y * y * y : (y - 16 / 116) / 7.787);
-	z = 1.08883 * ((z * z * z > 0.008856) ? z * z * z : (z - 16 / 116) / 7.787);
-
-	let r = x * 3.2406 + y * -1.5372 + z * -0.4986;
-	let g = x * -0.9689 + y * 1.8758 + z * 0.0415;
-	let b = x * 0.0557 + y * -0.2040 + z * 1.0570;
-
-	r = (r > 0.0031308) ? (1.055 * Math.pow(r, 1 / 2.4) - 0.055) : 12.92 * r;
-	g = (g > 0.0031308) ? (1.055 * Math.pow(g, 1 / 2.4) - 0.055) : 12.92 * g;
-	b = (b > 0.0031308) ? (1.055 * Math.pow(b, 1 / 2.4) - 0.055) : 12.92 * b;
-
-	return [
-		clamp(r, 0, 1) * 255,
-		clamp(g, 0, 1) * 255,
-		clamp(b, 0, 1) * 255,
-		alpha,
-	];
 }
 
 /* istanbul ignore next */